+++ src/job_widget.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "pch.h"
+#include "job_history.h"
 #include "ui_job_widget.h"
 
 class JobWidget : public QWidget {
@@ -14,6 +15,7 @@ public:
 
   void showDetails();
   bool wasSuccessful() const { return mSuccess; }
+  JobHistoryEntry historyEntry() const;
 
 public slots:
   void cancel();
@@ -28,8 +30,17 @@ private:
   bool mRunning = true;
   bool mSuccess = false;
   QProcess *mProcess;
+  QDateTime mStartedAt;
+  QDateTime mFinishedAt;
 
   QStringList mArgs;
+  QString mInfo;
+  QString mSource;
+  QString mDest;
+  qint64 mBytes = 0;
+  int mFiles = 0;
+  int mErrors = 0;
+  int mExitCode = 0;
   QHash<QString, QLabel *> mActive;
   QLabel *mOverflowLabel = nullptr;